+++ core/src/components/toggle/toggle.tsx
@@ -23,6 +23,7 @@ import type { ToggleChangeEventDetail } from './toggle-interface';
  *
  * @part track - The background track of the toggle.
  * @part handle - The toggle handle, or knob, used to change the checked state.
+ * @part label - The label text describing the toggle.
  */
 @Component({
   tag: 'ion-toggle',
@@ -223,6 +224,8 @@ export class Toggle implements ComponentInterface {
     if (this.legacyFormController.hasLegacyControl()) {
       this.ionStyle.emit({
         'interactive-disabled': this.disabled,
+        // TODO(FW-2990): remove this
+        legacy: !!this.legacy,
       });
     }
   }
@@ -375,6 +378,7 @@ export class Toggle implements ComponentInterface {
               'label-text-wrapper': true,
               'label-text-wrapper-hidden': !this.hasLabel,
             }}
+            part="label"
           >
             <slot></slot>
           </div>